qwen-messaging-agent/requirements.txt
torch>=2.1.0
transformers>=4.51.0
datasets>=2.14.0
peft>=0.7.0
bitsandbytes>=0.41.0
accelerate>=0.25.0
trl>=0.7.0
google-cloud-storage>=2.10.0
wandb>=0.15.0
flash-attn>=2.3.0
sentencepiece>=0.1.99
protobuf>=3.20.0

